add comment api, update category api

+++ backend/src/routers/category.router.ts
@@ -1,4 +1,5 @@
 import express from 'express'
+import ApiErrorResponse from '../utils/ApiErrorResponse'
 import Category from '../models/Category'
 
 export const categoryRouter = express.Router()
@@ -26,9 +27,13 @@ categoryRouter.post('/', express.json(), async (req, res) => {
   }
 })
 
-categoryRouter.post('/delete', express.json(), async (req, res) => {
+categoryRouter.post('/delete', express.json(), async (req, res, next) => {
   try {
     const { name } = req.body
+    const category = await Category.findOne({ name: name })
+    if (category.ideas.length > 0) {
+      return next(new ApiErrorResponse(`category 'name' are already attached in ideas`, 400))
+    }
     await Category.findOneAndDelete({ name })
     res.status(200).json({ success: 1 })
   } catch (err) {
